third_party/blink/web_tests/html5lib/generated/run-tests7-data-expected.txt

CONSOLE ERROR: Uncaught SyntaxError: Unexpected token '<'
../resources/tests7.dat: PASS
